% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/002_data_processing.R
\name{as.TR_data}
\alias{as.TR_data}
\title{Format tensor regression data}
\usage{
as.TR_data(y, X, eta = NULL)
}
\arguments{
\item{y}{a numeric response of length N}

\item{X}{an array in which the last margin is of size N}

\item{eta}{a matrix with N rows}
}
\value{
a list of class 'TR_data' in which dimensions have been verified as
compatible with the tensor regression functions
}
\description{
Format tensor regression data
}
