% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/904_plot_fns.R
\name{B_densities}
\alias{B_densities}
\title{Plot posterior densities of the tensor coefficient}
\usage{
B_densities(B, truth = NULL)
}
\arguments{
\item{B}{An array in which the last index corresponds to the \eqn{i}th draw
from the posterior distribution.}

\item{truth}{An optional array of true values for the tensor coefficient
which will be used to separate the density plots for voxels with true zero
values and true nonzero plots.}
}
\value{
A ggplot grob
}
\description{
Plot posterior densities of the tensor coefficient
}
