% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/905_BTR_CP_fns.R
\name{btr_compose_parafac}
\alias{btr_compose_parafac}
\title{Compose tensor from its parafac decomposition}
\usage{
btr_compose_parafac(betas)
}
\arguments{
\item{betas}{list of matrices}
}
\value{
array
}
\description{
Compose tensor from its parafac decomposition
}
