% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/403_BTR_CP.R
\name{BTR_CP}
\alias{BTR_CP}
\title{Bayesian tensor regression with the CP decomposition}
\usage{
BTR_CP(
  input,
  max_rank = 1,
  n_iter = 100,
  n_burn = 0,
  hyperparameters = NULL,
  save_dir = NULL
)
}
\arguments{
\item{input}{An object of class \code{TR_data} that contains (at least) the
elements \code{y} (a vector of response values) and \code{X} (an array of
covariate values). Optionally, \code{eta} (a matrix of nuisance covariates)
can also be included. Other list elements will be ignored.}

\item{max_rank}{(a scalar) The rank of the CP decomposition to be used}

\item{n_iter}{(a scalar) the number of posterior samples desired}

\item{n_burn}{(a scalar) the number of posterior samples to discard as a
burn-in}

\item{hyperparameters}{a list with the (scalar) elements \code{a.tau},
\code{b.tau}, \code{a.lambda}, \code{b.lambda}, \code{a.epsilon}, and
\code{b.epsilon} defining the values of the hyperparameters within the
model. If \code{NULL}, then default values will be used.}

\item{save_dir}{(a character) A path to a directory in which the temporary
results will be saved. Defaults to \code{NULL}. If \code{NULL}, then
temporary saves are not made.}
}
\value{
A list with the posterior samples
}
\description{
This code applies the model by Guhaniyogi et. al. (2017)
}
\examples{
\dontrun{
input <- TR_simulated_data()
results <- BTR_CP(input)
}
}
