% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/905_BTR_CP_fns.R
\name{btr_cp_all_B}
\alias{btr_cp_all_B}
\title{Wrapper to compose tensors from BTR_CP model object}
\usage{
btr_cp_all_B(btr_cp_object)
}
\arguments{
\item{btr_cp_object}{results from a call to BTR_CP}
}
\value{
list of arrays
}
\description{
Wrapper to compose tensors from BTR_CP model object
}
