% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/905_BTR_CP_fns.R
\name{btr_cp_final_B}
\alias{btr_cp_final_B}
\title{Produce a final estimate of the BTR CP tensor coefficient}
\usage{
btr_cp_final_B(btr_cp_object)
}
\arguments{
\item{btr_cp_object}{an object of class \code{BTRCP_result}}
}
\value{
an array of the final tensor estimate found using sequential 2-means
}
\description{
Produce a final estimate of the BTR CP tensor coefficient
}
\examples{
\dontrun{
input <- TR_simulated_data()
result <- BTR_CP(input)
final_B <- btr_cp_final_B(result)
}
}
