% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/905_BTR_CP_fns.R
\name{btr_cp_llik}
\alias{btr_cp_llik}
\title{Calculate log-likelihood from result object}
\usage{
btr_cp_llik(btr_object, input)
}
\arguments{
\item{btr_object}{result from BTR_CP}

\item{input}{input list used to run BTR_CP}
}
\value{
vector of log-likelihoods
}
\description{
Calculate log-likelihood from result object
}
