% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/900_BTRR_sampling_fns.R
\name{BTRR_draw_beta}
\alias{BTRR_draw_beta}
\title{Draw betas}
\usage{
BTRR_draw_beta(Y, x, betas, Sigma_AR, tau, Phi, W, j, r)
}
\arguments{
\item{Y}{Array of the response values}

\item{x}{Matrix of the covariates}

\item{betas, W}{Lists of length \eqn{D} comprised of \eqn{p_j} by \eqn{R}
matrices}

\item{Sigma_AR}{A covariance matrix based off of the observational variance \
and the covariance structure.}

\item{tau}{A scalar for the global variance}

\item{Phi}{A vector for the rank-weighting of the variance}

\item{j}{Dimension index}

\item{r}{Rank index}
}
\value{
A list of length \eqn{D} made up of matrices with dimensions
  \eqn{p_j} by \eqn{R}.
}
\description{
Draw beta variables from their posterior full conditional distributions. As a
  note, this is done with a loop in the code for the Markov Chain Monte
  Carlo.
}
\keyword{internal}
