% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/900_BTRR_sampling_fns.R
\name{BTRR_draw_Xi}
\alias{BTRR_draw_Xi}
\title{Draw Xi}
\usage{
BTRR_draw_Xi(Xi, cov_Metro, betas, W, tau, a)
}
\arguments{
\item{Xi}{The previous iteration values for Xi}

\item{cov_Metro}{The Metropolis proposal covariance}

\item{betas, W}{Lists of length \eqn{D} comprised of \eqn{p_j} by \eqn{R}
matrices}

\item{tau}{A scalar value for the global variance}

\item{a}{A scalar value for the value of variance weight}
}
\value{
A vector of length \eqn{R - 1}
}
\description{
Perform a Metropolis-Hastings draw of the Xi parameters.
}
\keyword{internal}
