% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/903_post_hoc_fns.R
\name{BTRR_extract_B}
\alias{BTRR_extract_B}
\title{Construct tensor coefficient from results}
\usage{
BTRR_extract_B(btrr_result)
}
\arguments{
\item{btrr_result}{A list object with class \code{BTRR_result}}
}
\value{
An array in which the last index corresponds to the \eqn{i}th draw
  from the posterior distribution.
}
\description{
As the \code{BTRR_single_subject} function saves posterior samples of the
  tensor decomposition components, the tensor coefficient is composed from
  results of class \code{BTRR_result}.
}
