% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/901_BTRR_GGM_sampling_fns.R
\name{BTRR_GGM_BtWB}
\alias{BTRR_GGM_BtWB}
\title{Quadratic calculation}
\usage{
BTRR_GGM_BtWB(beta_g, W_g)
}
\arguments{
\item{beta_g}{region beta}

\item{W_g}{region omega}
}
\value{
The results of the quadratic (D x Rank)
}
\description{
Quadratic calculation
}
\keyword{internal}
