% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/901_BTRR_GGM_sampling_fns.R
\name{BTRR_GGM_check_data}
\alias{BTRR_GGM_check_data}
\title{Check that the input data are compatible with the \code{BTRR_GGM} model function.}
\usage{
BTRR_GGM_check_data(input)
}
\arguments{
\item{input}{an input object to be checked}
}
\value{
If the input object passes all of the checks, then it is also returned.
}
\description{
Check that the input data are compatible with the \code{BTRR_GGM} model function.
}
\keyword{internal}
