% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/901_BTRR_GGM_sampling_fns.R
\name{BTRR_GGM_draw_beta}
\alias{BTRR_GGM_draw_beta}
\title{Draw betas}
\usage{
BTRR_GGM_draw_beta(Y_g, x, d_g, beta_g, sig2y, tau.g, phi.g, omega.g, j, r)
}
\arguments{
\item{Y_g}{region response}

\item{x}{covariate}

\item{d_g}{region effect}

\item{beta_g}{activation effect}

\item{sig2y}{observation error}

\item{tau.g}{global variance parameter}

\item{phi.g}{local rank variance weighting}

\item{omega.g}{local variance}

\item{j}{dimension margin}

\item{r}{rank}
}
\value{
vector beta updates
}
\description{
Draw betas
}
\keyword{internal}
