% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/901_BTRR_GGM_sampling_fns.R
\name{BTRR_GGM_draw_d}
\alias{BTRR_GGM_draw_d}
\title{Draw d}
\usage{
BTRR_GGM_draw_d(input, params, Sig, sig2y)
}
\arguments{
\item{input}{the input data}

\item{params}{the list of parallel results}

\item{Sig}{The interregion precision}

\item{sig2y}{The observation variance}
}
\value{
the matrix d
}
\description{
Draw d
}
\keyword{internal}
