% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/901_BTRR_GGM_sampling_fns.R
\name{BTRR_GGM_draw_lambda}
\alias{BTRR_GGM_draw_lambda}
\title{Update lambda}
\usage{
BTRR_GGM_draw_lambda(a.lambda, b.lambda, beta_g, phi.g, tau.g)
}
\arguments{
\item{a.lambda}{hyperparameter}

\item{b.lambda}{hyperparameter}

\item{beta_g}{region beta}

\item{phi.g}{region variance weights}

\item{tau.g}{region global variance}
}
\value{
update for lambda
}
\description{
Update lambda
}
\keyword{internal}
