% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/901_BTRR_GGM_sampling_fns.R
\name{BTRR_GGM_draw_omega}
\alias{BTRR_GGM_draw_omega}
\title{Update local region variance parameters}
\usage{
BTRR_GGM_draw_omega(beta_g, tau.g, phi.g, lambda.g)
}
\arguments{
\item{beta_g}{region location parameters}

\item{tau.g}{region global variance}

\item{phi.g}{region covariance weights}

\item{lambda.g}{region covariance shrinkage parameter}
}
\value{
an update for omega
}
\description{
Update local region variance parameters
}
\keyword{internal}
