% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/901_BTRR_GGM_sampling_fns.R
\name{BTRR_GGM_draw_Phi}
\alias{BTRR_GGM_draw_Phi}
\title{Draw Phi_g under the stick-breaking prior}
\usage{
BTRR_GGM_draw_Phi(BtWB_g, Xi_g, accept_g, cov_Metro_g, p_g, tau_g, alpha_g)
}
\arguments{
\item{BtWB_g}{region normalized beta quadratic}

\item{Xi_g}{last draw region Xi}

\item{accept_g}{acceptance count for region g}

\item{cov_Metro_g}{metropolis covariance for region g}

\item{tau_g}{region tau}

\item{alpha}{region alpha}
}
\value{
List with a Phi update,  the acceptance number update, and the Xi
update
}
\description{
Draw Phi_g under the stick-breaking prior
}
\keyword{internal}
