% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/901_BTRR_GGM_sampling_fns.R
\name{BTRR_GGM_draw_Sig_Upsilon}
\alias{BTRR_GGM_draw_Sig_Upsilon}
\title{Draw Sigma and Upsilon}
\usage{
BTRR_GGM_draw_Sig_Upsilon(d, g, zeta, Sig, Upsilon)
}
\arguments{
\item{d}{subject-region effects}

\item{g}{region}

\item{zeta}{scalar parameter}

\item{Sig}{Precision}

\item{Upsilon}{Precision parameter}
}
\value{
Updated Sig and Upsilon
}
\description{
Draw Sigma and Upsilon
}
\keyword{internal}
