% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/901_BTRR_GGM_sampling_fns.R
\name{BTRR_GGM_draw_tau}
\alias{BTRR_GGM_draw_tau}
\title{Update tau}
\usage{
BTRR_GGM_draw_tau(a.tau, b.tau, p_g, phi.g, BtWB_g)
}
\arguments{
\item{a.tau}{hyperparameter}

\item{b.tau}{hyperparameter}

\item{p_g}{region dimension}

\item{phi.g}{region rank variation weights}

\item{BtWB_g}{quadratic}
}
\value{
update for tau
}
\description{
Update tau
}
\keyword{internal}
