% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/901_BTRR_GGM_sampling_fns.R
\name{BTRR_GGM_empty_results}
\alias{BTRR_GGM_empty_results}
\title{Create a preallocated result list object}
\usage{
BTRR_GGM_empty_results(p, n, rank, n_iter)
}
\arguments{
\item{p}{a list of length equal to the number of regions giving the response
tensor dimensions for that region}

\item{rank}{number of ranks in the model}

\item{n_iter}{number of iterations for the MCMC}
}
\value{
preallocated result list object
}
\description{
Create a preallocated result list object
}
\keyword{internal}
