% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/901_BTRR_GGM_sampling_fns.R
\name{BTRR_GGM_griddy_Gibbs_alpha}
\alias{BTRR_GGM_griddy_Gibbs_alpha}
\title{Griddy Gibbs draw for alpha}
\usage{
BTRR_GGM_griddy_Gibbs_alpha(
  g,
  M,
  p,
  TT,
  Rank,
  cov_Metro,
  alpha.g,
  betas,
  W,
  Xi,
  tau,
  a.tau,
  b.tau,
  d,
  input,
  sig2y
)
}
\arguments{
\item{g}{region index}

\item{M}{number of samples per proposal}

\item{p}{region tensor dimensions}

\item{TT}{number of time steps}

\item{Rank}{model rank}

\item{cov_Metro}{Metropolis covariance}

\item{alpha.g}{grid of possible values for alpha}

\item{betas}{tensor decomposition components}

\item{W}{beta individual variances}

\item{Xi}{variance rank weight stick breaks}

\item{tau}{region-wide variance}

\item{a.tau}{hyperparameter}

\item{b.tau}{hyperparameter}

\item{d}{region-subject intercepts}

\item{input}{input data}

\item{sig2y}{observation variance}
}
\value{
a scalar value for alpha
}
\description{
Griddy Gibbs draw for alpha
}
\keyword{internal}
