% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{btrr_ggm_result}
\alias{btrr_ggm_result}
\title{Results from the simulated sample Bayesian Tensor Response Regression with Gaussian Graphical Model}
\format{
A BTRR_GGM_result object
}
\usage{
data(btrr_ggm_result)
}
\description{
This is a \code{BTRR_GMM_result} object from running the BTRR_GGM function on
  the simulated data created in the vignette.
}
\keyword{datasets}
