% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/903_post_hoc_fns.R
\name{BTRR_post_hoc_AR1_llik}
\alias{BTRR_post_hoc_AR1_llik}
\title{Calculate the log-likelihood}
\usage{
BTRR_post_hoc_AR1_llik(input_data, results_obj)
}
\arguments{
\item{input_data}{A list of class \code{TRR_data} containing (at least) a
response array \code{Y} and a design matrix \code{x}.}

\item{results_obj}{A list of class \code{BTRR_result} containing the draws
from the posterior distribution from \code{\link{BTRR_single_subject}}.}
}
\value{
A vector with length equal to the number of samples from the
  posterior distribution containing the log-likelihood values for each sample.
}
\description{
Calculate the log-likelihood for the Bayesian tensor response regression
  model. This can be useful when performing diagnostics. This is excluded
  from the Markov Chain function for performance gains.
}
