% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BTRR_results}
\alias{BTRR_results}
\title{Results from the simulated sample Bayesian Tensor Response Regression}
\format{
A list of three \code{BTRR_result} objects.
}
\usage{
BTRR_results
}
\description{
This object is a list containing the results of the
\code{BTRR_single_subject} function for ranks 1, 2, and 3 based on the
simulated data created in the vignette.
}
\keyword{datasets}
