% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/903_post_hoc_fns.R
\name{BTRT_all_B}
\alias{BTRT_all_B}
\title{Compose all tensors with Tucker decomposition}
\usage{
BTRT_all_B(btrt_object)
}
\arguments{
\item{btrt_object}{Output object from \code{BTRTucker}}
}
\value{
An array of samples from the posterior distribution of the tensor
coefficient. The last dimension corresponds to the draw number
}
\description{
Uses the tensor decomposition from Tucker (1927) to compose a tensor
coefficient draw from a \code{BTRT_result} object.
}
\examples{
\dontrun{
input <- TR_simulated_data()
results <- BTRTucker(input)
all_B <- BTRT_all_B(results)
}
}
