% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/902_BTRTucker_sampling_fns.R
\name{BTRT_draw_gam}
\alias{BTRT_draw_gam}
\title{Draw gamma from PFC}
\usage{
BTRT_draw_gam(eta, Sig_0, mu_gam, y_til, sig_y2)
}
\arguments{
\item{eta}{nuisance input}

\item{Sig_0}{prior covariance}

\item{mu_gam}{prior mean}

\item{y_til}{response minus BX}

\item{sig_y2}{observational variance}
}
\value{
gamma update
}
\description{
Draw gamma from PFC
}
\keyword{internal}
