% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/902_BTRTucker_sampling_fns.R
\name{BTRT_draw_lam}
\alias{BTRT_draw_lam}
\title{Update local shrinkage}
\usage{
BTRT_draw_lam(a_lam, b_lam, betas, tau)
}
\arguments{
\item{a_lam}{hyperparameter value}

\item{b_lam}{hyperparameter value}

\item{betas}{tensor decomposition matrices}

\item{tau}{global variance}
}
\value{
update to local shrinkage
}
\description{
Update local shrinkage
}
\keyword{internal}
