% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/902_BTRTucker_sampling_fns.R
\name{BTRT_draw_omega}
\alias{BTRT_draw_omega}
\title{Draw local variance}
\usage{
BTRT_draw_omega(lam, betas, tau)
}
\arguments{
\item{lam}{local shrinkage parameter}

\item{betas}{tensor decomposition matrices}

\item{tau}{global variance}
}
\value{
update to local variance
}
\description{
Draw local variance
}
\keyword{internal}
