% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/902_BTRTucker_sampling_fns.R
\name{BTRT_draw_sig_y2}
\alias{BTRT_draw_sig_y2}
\title{Update observational variance}
\usage{
BTRT_draw_sig_y2(a.sig, b.sig, y_til)
}
\arguments{
\item{a.sig}{hyperparameter}

\item{b.sig}{hyperparameter}

\item{y_til}{response less effects from covariates}
}
\value{
update to observational variance
}
\description{
Update observational variance
}
\keyword{internal}
