% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/902_BTRTucker_sampling_fns.R
\name{BTRT_draw_tau}
\alias{BTRT_draw_tau}
\title{Update global variance}
\usage{
BTRT_draw_tau(a_tau, b_tau, betas, W)
}
\arguments{
\item{a_tau}{hyperparameter}

\item{b_tau}{hyperparameter}

\item{betas}{tensor decomposition matrices}

\item{W}{local variance}
}
\value{
update to local variance
}
\description{
Update global variance
}
\keyword{internal}
