% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/902_BTRTucker_sampling_fns.R
\name{BTRT_draw_U}
\alias{BTRT_draw_U}
\title{Update core tensor shrinkage parameter}
\usage{
BTRT_draw_U(a.u, b.u, G, z)
}
\arguments{
\item{a.u}{hyperparameter}

\item{b.u}{hyperparameter}

\item{G}{core tensor}

\item{z}{global variance}
}
\value{
update to core tensor shrinkage
}
\description{
Update core tensor shrinkage parameter
}
\keyword{internal}
