% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/902_BTRTucker_sampling_fns.R
\name{BTRT_draw_V}
\alias{BTRT_draw_V}
\title{Update to core tensor local variance}
\usage{
BTRT_draw_V(U, z, G)
}
\arguments{
\item{U}{local shrinkage parameters}

\item{z}{global shrinkage parameters}

\item{G}{core tensor}
}
\value{
update to core tensor local variance
}
\description{
Update to core tensor local variance
}
\keyword{internal}
