% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/902_BTRTucker_sampling_fns.R
\name{BTRT_draw_z}
\alias{BTRT_draw_z}
\title{Update core tensor global variance}
\usage{
BTRT_draw_z(a.z, b.z, G, V)
}
\arguments{
\item{a.z}{hyperparameter}

\item{b.z}{hyperparameter}

\item{G}{core tensor}

\item{V}{core tensor local variance}
}
\value{
update to core tensor global variance
}
\description{
Update core tensor global variance
}
\keyword{internal}
