% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/903_post_hoc_fns.R
\name{BTRT_estimate_B}
\alias{BTRT_estimate_B}
\title{Produce a final tensor coefficient estimate}
\usage{
BTRT_estimate_B(all_B)
}
\arguments{
\item{all_B}{An array of tensor coefficient posterior draws in which the
final dimension corresponds to the posterior sample index.}
}
\value{
An array as a final estimate of the tensor coefficient
}
\description{
This uses the sequential 2-means post-hoc variable-selection technique from
  Li and Pati (2018) to produce a point estimate of the tensor coefficient
  from all of posterior samples of the tensor coefficient.
}
\examples{
\dontrun{
input <- TR_simulated_data()
results <- BTRTucker(input)
all_B <- BTRT_all_B(results)
estimate_B <- BTRT_estimate_B(all_B)
}
}
