% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/903_post_hoc_fns.R
\name{BTRT_final_B}
\alias{BTRT_final_B}
\title{Produce a final estimate of the tensor coefficient}
\usage{
BTRT_final_B(btrt_object)
}
\arguments{
\item{btrt_object}{An object of class \code{BTRT_result} output from
\code{BTRTucker}}
}
\value{
An array estimate of the tensor coefficient
}
\description{
This is a helper function that combines \code{BTRT_all_B},
\code{BTRT_estimate_B}, \code{composeTuckerCore}, \code{s2m_B}, and
\code{s2m}.
}
\examples{
\dontrun{
input <- TR_simulated_data()
result <- BTRTucker(input)
final_B <- BTRT_final_B(result)
}
}
