% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/402_BTRTucker.R
\name{BTRTucker}
\alias{BTRTucker}
\title{Bayesian tensor regression with the Tucker decomposition}
\usage{
BTRTucker(
  input,
  ranks = rep(1, length(dim(input$X)) - 1),
  n_iter = 100,
  n_burn = 0,
  hyperparameters = NULL,
  save_dir = NULL
)
}
\arguments{
\item{input}{An object of class \code{TR_data} that contains (at least) the
elements \code{y} (a vector of response values) and \code{X} (an array of
covariate values). Optionally, \code{eta} (a matrix of nuisance covariates)
can also be included. Other list elements will be ignored.}

\item{ranks}{A vector of length \code{length(dim(input$X)) - 1} giving the
desired number of ranks per dimension of the tensor coefficient}

\item{n_iter}{(a scalar) the number of posterior samples desired}

\item{n_burn}{(a scalar) the number of posterior samples to discard as a
burn-in}

\item{hyperparameters}{a list with the (scalar) elements \code{a.tau},
\code{b.tau}, \code{a.lam}, \code{b.lam}, \code{nu}, \code{s_02},
\code{a.sig}, \code{b.sig}, \code{Sig_0}, \code{mu_gam}, \code{alpha.grid},
\code{a.u}, \code{b.u}, \code{a.z}, and/or \code{b.z},
defining the values of the hyperparameters within the
model. If \code{NULL}, then default values will be used. It is also
possible to specify only a subset of the hyperparameters. The
remaining hyperparameters are set to their default values.}

\item{save_dir}{(a character) A path to a directory in which the temporary
results will be saved. Defaults to the current working directory. If
\code{NULL}, no temporary saves are made.}
}
\value{
A list with the posterior samples
}
\description{
Bayesian tensor regression with the Tucker decomposition
}
\examples{
\dontrun{
input <- TR_simulated_data()
results <- BTRTucker(input)
}
}
