% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/906_ftr_functions.R
\name{compose_parafac}
\alias{compose_parafac}
\title{Compose a tensor from its PARAFAC decomposition}
\usage{
compose_parafac(betas)
}
\arguments{
\item{betas}{a list, where the first level indexes dimension, and the columns
of the matrices within each list represent the rank}
}
\value{
an array
}
\description{
Compose a tensor from its PARAFAC decomposition
}
\keyword{internal}
