% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/906_ftr_functions.R
\name{compose_tucker_ftr}
\alias{compose_tucker_ftr}
\title{Compose a tensor from its Tucker decomposition}
\usage{
compose_tucker_ftr(bb, gg)
}
\arguments{
\item{bb}{list}

\item{gg}{array}
}
\value{
array
}
\description{
Compose a tensor from its Tucker decomposition
}
\keyword{internal}
