% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/906_ftr_functions.R
\name{compose_tucker_ftr_vec}
\alias{compose_tucker_ftr_vec}
\title{Compose a tensor from the Tucker decomposition and vectorize}
\usage{
compose_tucker_ftr_vec(bb, gg)
}
\arguments{
\item{bb}{list}

\item{gg}{array}
}
\value{
vector
}
\description{
Compose a tensor from the Tucker decomposition and vectorize
}
\keyword{internal}
