% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/900_BTRR_sampling_fns.R
\name{composeParafac}
\alias{composeParafac}
\title{Compose a tensor from its CANDECOMP/PARAFAC (CP) decomposition}
\usage{
composeParafac(bb)
}
\arguments{
\item{bb}{A list of length D in which each element is a p_d by R matrix}
}
\value{
A single array-class tensor. In two-dimensions, this will be returned
  as a matrix.
}
\description{
This function takes a list of length D containing all of
the components of the CP decomposition and returning a D-dimensional
tensor.
}
