% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/903_post_hoc_fns.R
\name{composeTuckerCore}
\alias{composeTuckerCore}
\title{Compose a tensor based on the Tucker Decomposition}
\usage{
composeTuckerCore(bb, gg)
}
\arguments{
\item{bb}{a list of tensor decomposition matrices in which the number of
columns corresponds to the dimension margins}

\item{gg}{an array in which the length of each dimension margin corresponds
to the number of ranks to decompose each dimension}
}
\value{
an array (tensor) composed using the Tucker tensor decomposition
}
\description{
Compose a tensor based on the Tucker Decomposition
}
