% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/905_BTR_CP_fns.R
\name{cp_draw_betas}
\alias{cp_draw_betas}
\title{Draw betas from their posterior FCs}
\usage{
cp_draw_betas(X, betas, sig_y2, y, gam, eta, W, Phi, tau, j, r)
}
\arguments{
\item{X}{design array}

\item{betas}{list of matrices}

\item{sig_y2}{scalar}

\item{y}{vector}

\item{gam}{vector}

\item{eta}{design matrix}

\item{W}{list of matrices}

\item{Phi}{vector or matrix}

\item{tau}{scalar}

\item{j}{(scalar) dimension index}

\item{r}{(scalar) rank index}
}
\value{
vector
}
\description{
Draw betas from their posterior FCs
}
\keyword{internal}
