% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/905_BTR_CP_fns.R
\name{cp_draw_gam}
\alias{cp_draw_gam}
\title{Draw gam from posterior FC}
\usage{
cp_draw_gam(eta, Sig_0, y_til, sig_y2)
}
\arguments{
\item{eta}{design matrix}

\item{Sig_0}{matrix}

\item{y_til}{vector}

\item{sig_y2}{scalar}
}
\value{
vector
}
\description{
Draw gam from posterior FC
}
\keyword{internal}
