% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/905_BTR_CP_fns.R
\name{cp_draw_gam_two}
\alias{cp_draw_gam_two}
\title{Second way to draw gam from posterior FC}
\usage{
cp_draw_gam_two(eta, Sig_0, mu_gam, y_til, sig_y2)
}
\arguments{
\item{eta}{design matrix}

\item{Sig_0}{matrix}

\item{mu_gam}{vector}

\item{y_til}{vector}

\item{sig_y2}{scalar}
}
\value{
vector
}
\description{
Second way to draw gam from posterior FC
}
\keyword{internal}
