% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/905_BTR_CP_fns.R
\name{cp_draw_lam}
\alias{cp_draw_lam}
\title{Draw lam from posterior FC}
\usage{
cp_draw_lam(a_lam, b_lam, betas, Phi, tau)
}
\arguments{
\item{a_lam}{scalar}

\item{b_lam}{scalar}

\item{betas}{list of matrices}

\item{Phi}{vector or matrix}

\item{tau}{scalar}
}
\value{
vector
}
\description{
Draw lam from posterior FC
}
\keyword{internal}
