% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/905_BTR_CP_fns.R
\name{cp_draw_Phi}
\alias{cp_draw_Phi}
\title{Draw Phi variable from the posterior full conditional}
\usage{
cp_draw_Phi(n, alpha, betas, W, b_tau)
}
\arguments{
\item{n}{scalar}

\item{alpha}{scalar}

\item{betas}{list of matrices}

\item{W}{list of matrices}

\item{b_tau}{scalar}
}
\value{
vector or matrix of draws from the posterior full conditional
}
\description{
Draw Phi variable from the posterior full conditional
}
\keyword{internal}
