% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/905_BTR_CP_fns.R
\name{cp_draw_sig_y2}
\alias{cp_draw_sig_y2}
\title{Draw sig_y2 from posterior FC}
\usage{
cp_draw_sig_y2(nu, s_02, y_til, mu_gam, eta)
}
\arguments{
\item{nu}{scalar}

\item{s_02}{scalar}

\item{y_til}{vector}

\item{mu_gam}{vector}

\item{eta}{design matrix}
}
\value{
scalar
}
\description{
Draw sig_y2 from posterior FC
}
\keyword{internal}
