% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/905_BTR_CP_fns.R
\name{cp_draw_sig_y2_two}
\alias{cp_draw_sig_y2_two}
\title{Second way to draw sig_y2 from posterior FC}
\usage{
cp_draw_sig_y2_two(a.sig, b.sig, y.til)
}
\arguments{
\item{a.sig}{scalar}

\item{b.sig}{scalar}

\item{y.til}{vector}
}
\value{
scalar
}
\description{
Second way to draw sig_y2 from posterior FC
}
\keyword{internal}
