% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/905_BTR_CP_fns.R
\name{cp_draw_tau}
\alias{cp_draw_tau}
\title{Draw tau from the posterior full conditional}
\usage{
cp_draw_tau(n, a_tau, b_tau, betas, W, Phi)
}
\arguments{
\item{n}{scalar}

\item{a_tau}{scalar}

\item{b_tau}{scalar}

\item{betas}{list of matrices}

\item{W}{list of matrices}

\item{Phi}{vector or matrix}
}
\value{
vector
}
\description{
Draw tau from the posterior full conditional
}
\keyword{internal}
