% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/905_BTR_CP_fns.R
\name{cp_draw_W}
\alias{cp_draw_W}
\title{Draw W from posterior FC}
\usage{
cp_draw_W(lam, betas, Phi, tau)
}
\arguments{
\item{lam}{vector}

\item{betas}{list of matrices}

\item{Phi}{vector or matrix}

\item{tau}{scalar}
}
\value{
list of matrices
}
\description{
Draw W from posterior FC
}
\keyword{internal}
