% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/900_BTRR_sampling_fns.R
\name{draw_k}
\alias{draw_k}
\title{Draw k}
\usage{
draw_k(Y, x, betas, sigma_epsilon_sq)
}
\arguments{
\item{Y}{Array of the response values}

\item{x}{Matrix of the covariates}

\item{betas}{List of length \eqn{D} comprised of \eqn{p_j} by \eqn{R}
matrices}

\item{sigma_epsilon_sq}{Scalar observational variance}
}
\value{
A scalar
}
\description{
Draw k from its posterior full conditional distribution.
}
\keyword{internal}
