% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/900_BTRR_sampling_fns.R
\name{draw_tau}
\alias{draw_tau}
\title{Draw tau}
\usage{
draw_tau(a.tau, b.tau, betas, W, Phi)
}
\arguments{
\item{a.tau, b.tau}{Scalar values for the hyperparameters}

\item{betas, W}{Lists of length \eqn{D} comprised of \eqn{p_j} by \eqn{R}
matrices}

\item{Phi}{A vector whose elements sum to 1}
}
\value{
A scalar
}
\description{
Draw a value of tau using its posterior full conditional distribution.
}
\keyword{internal}
