% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/906_ftr_functions.R
\name{ftr_log_likelihood}
\alias{ftr_log_likelihood}
\title{Obtain the log-likelihood for the tensor regression linear model}
\usage{
ftr_log_likelihood(input, B, gam)
}
\arguments{
\item{input}{An object of class \code{TR_data} that contains (at least) the
elements \code{y} (a vector of response values) and \code{X} (an array of
covariate values). Optionally, \code{eta} (a matrix of nuisance covariates)
can also be included. Other list elements will be ignored.}

\item{B}{tensor coefficient}

\item{gam}{vector coefficients}
}
\value{
scalar
}
\description{
Obtain the log-likelihood for the tensor regression linear model
}
\keyword{internal}
