% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/906_ftr_functions.R
\name{ftr_model_BIC}
\alias{ftr_model_BIC}
\title{Get the BIC for an FTR model fit}
\usage{
ftr_model_BIC(ftr_object, input)
}
\arguments{
\item{ftr_object}{The result from a call to FTR_CP or FTRTucker}

\item{input}{list with elements \code{y} and \code{X}}
}
\value{
scalar
}
\description{
Get the BIC for an FTR model fit
}
\keyword{internal}
