% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/904_plot_fns.R
\name{gg_tile_plot}
\alias{gg_tile_plot}
\title{Make a tile plot using ggplot2}
\usage{
gg_tile_plot(x, title = NULL)
}
\arguments{
\item{x}{a matrix}

\item{title}{the plot title}
}
\value{
a ggplot grob
}
\description{
Make a tile plot using ggplot2
}
