% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/904_plot_fns.R
\name{ggplot_ts}
\alias{ggplot_ts}
\title{Time series ggplot}
\usage{
ggplot_ts(x, title = NULL)
}
\arguments{
\item{x}{A numeric vector}

\item{title}{A character string of the plot title}
}
\value{
A ggplot grob
}
\description{
Time series ggplot
}
