% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/902_BTRTucker_sampling_fns.R
\name{kFold}
\alias{kFold}
\title{Perform mode-k matricization on an array}
\usage{
kFold(x, k)
}
\arguments{
\item{x}{an array}

\item{k}{an integer or vector of integers}
}
\value{
a matrix with dimensions \code{prod(dim(x)[k]) x prod(dim(X)[-k])}
}
\description{
Perform mode-k matricization on an array
}
\examples{
set.seed(47408)
p <- rep(50,3)
N <- 20
A <- array(rnorm(prod(p,N)),dim = c(p,N))
Ak <- kFold(A,2)
}
