% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/906_ftr_functions.R
\name{khatri_rao}
\alias{khatri_rao}
\title{Khatri Rao product}
\usage{
khatri_rao(A, B = NULL)
}
\arguments{
\item{A}{matrix}

\item{B}{matrix}
}
\value{
Khatri Rao matrix product
}
\description{
Khatri Rao product
}
\keyword{internal}
