% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/904_plot_fns.R
\name{melt_array}
\alias{melt_array}
\title{Melt and array into a data frame}
\usage{
melt_array(x, value.name = "value")
}
\arguments{
\item{x}{an array or matrix}

\item{value.name}{an optional input for the column name containing data}
}
\value{
a data frame with the melted data
}
\description{
This is a replacement for reshape2::melt, intended to reduce the number of
package dependencies
}
\examples{
A <- array(1:27, dim = c(3,3,3))
melt_array(A)
}
