% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/904_plot_fns.R
\name{melt_list}
\alias{melt_list}
\title{Wrapper for melting arrays within lists}
\usage{
melt_list(x)
}
\arguments{
\item{x}{a list containing matrices or arrays}
}
\value{
a data frame with the melted data
}
\description{
Wrapper for melting arrays within lists
}
\examples{
A <- array(1:8, dim = c(2,2,2))
B <- rep(list(A),2)
melt_list(B)
}
