% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/900_BTRR_sampling_fns.R
\name{mode_k_matriz}
\alias{mode_k_matriz}
\title{Matricization of a tensor}
\usage{
mode_k_matriz(tens, k)
}
\arguments{
\item{tens}{An array of dimension \eqn{D}}

\item{k}{An integer less than or equal to \eqn{D}, which will become the
first index in the matricized array}
}
\value{
A matrix
}
\description{
Matricization of a tensor
}
\keyword{internal}
