% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/900_BTRR_sampling_fns.R
\name{s2m}
\alias{s2m}
\title{Sequential 2-means number of nonzero-values parameters for an iteration}
\usage{
s2m(x, b)
}
\arguments{
\item{x}{A vector of parameters from a single draw from a posterior
distribution}

\item{b}{The maximum distance at which two clusters of parameters are
considered equal}
}
\value{
A scalar representing the number of nonzero-valued parameters in a
  particular draw from a posterior distribution
}
\description{
Sequential 2-means number of nonzero-values parameters for an iteration
}
\examples{
post_sample <- numeric(100)
for(k in 1:100) {
  g <- runif(1)
  if(g < 0.8) {
    post_sample[k] <- rnorm(1,0,0.2)
  } else {
    post_sample[k] <- rnorm(1,0,0.2)
  }
}
how_many_nonzero <- s2m(post_sample, b = sd(post_sample))
}
