% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/900_BTRR_sampling_fns.R
\name{stick_break_log_posterior}
\alias{stick_break_log_posterior}
\title{The log of the posterior probability for the stick-breaking prior structure}
\usage{
stick_break_log_posterior(Xi, betas, W, tau, a)
}
\arguments{
\item{Xi}{A vector of length \eqn{R - 1}}

\item{betas, W}{Lists of length \eqn{D} comprised of \eqn{p_j} by \eqn{R}
matrices}

\item{tau}{A scalar value for the global variance}

\item{a}{A scalar value for the value of variance weight}
}
\value{
An integer value
}
\description{
The log of the posterior probability for the stick-breaking prior structure
}
\keyword{internal}
