% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/900_BTRR_sampling_fns.R
\name{stick_values}
\alias{stick_values}
\title{Calculate stick values}
\usage{
stick_values(breaks)
}
\arguments{
\item{breaks}{A vector of length \eqn{n-1} of break proportions}
}
\value{
A vector of length \eqn{n} of break locations between 0 and 1
}
\description{
Calculate stick values
}
\keyword{internal}
