% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/001_data_simulation.R
\name{TR_simulated_data}
\alias{TR_simulated_data}
\title{Simulated tensor regression data}
\usage{
TR_simulated_data(
  subjects = 400,
  tensor_dims = c(50, 50),
  CNR = 1,
  num_active = 3,
  other_covar = c(1, 1)
)
}
\arguments{
\item{subjects}{A positive integer indicating the number of subjects for
which to generate data}

\item{tensor_dims}{A vector that indicates the dimensions of the tensor
covariate for each subject}

\item{CNR}{The contrast-to-noise ratio, defined by Welvaert and Rosseel
(2013) as the size of the (largest) true coefficient value divided by
the observational noise.}

\item{num_active}{The number of contiguous nonzero activation regions}

\item{other_covar}{A vector of true values for other (nuisance) coefficients
that will be used to generate non-tensor-valued covariates}
}
\value{
An object of class \code{TR_data} that contains the
elements \code{y} (a vector of response values), \code{X} (an array of
covariate values), \code{eta} (a matrix of nuisance covariates),
\code{true_B} (an array with the true tensor coefficient values), and
\code{gam} (a vector with the true nuisance coefficient values)
}
\description{
Simulated tensor regression data
}
\examples{
input <- TR_simulated_data()
}
