% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/001_data_simulation.R
\name{TRR_GGM_simulated_data}
\alias{TRR_GGM_simulated_data}
\title{Simulated data for tensor response regression with Gaussian graphical model}
\usage{
TRR_GGM_simulated_data(
  subjects = 20,
  regions = 10,
  max_time = 200,
  margin_sizes = c(10, 10, 10),
  SNR = 1,
  CNR = 0.05,
  conn_regions = 2,
  conn_level = 0.9
)
}
\arguments{
\item{subjects}{The number of subjects in the data (a scalar)}

\item{regions}{The number of response tensors per subject in the simulated
data}

\item{max_time}{The length of the tensor time series (a scalar)}

\item{margin_sizes}{The sizes of the tensor coefficients (a vector)}

\item{SNR}{The signal-to-noise ratio (a scalar), as defined in
Welvaert, M., & Rosseel, Y. (2013). On the definition of signal-to-noise
ratio and contrast-to-noise ratio for fMRI data. PloS one, 8(11), e77089.}

\item{CNR}{The contrast-to-noise ratio (a scalar), as defined in
Welvaert, M., & Rosseel, Y. (2013). On the definition of signal-to-noise
ratio and contrast-to-noise ratio for fMRI data. PloS one, 8(11), e77089.}

\item{conn_regions}{The number of response tensors that should have nonzero
partial correlations}

\item{conn_level}{The partial correlation for the connected regions}
}
\value{
A list with class \code{TRR_GGM_data} with elements \code{Y},
  \code{x}, \code{true_d}, \code{true_B}, and \code{true_d_covar}
}
\description{
Simulated data for tensor response regression with Gaussian graphical model
}
\examples{
\dontrun{
input <- TRR_GGM_simulated_data()
}
}
