% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/001_data_simulation.R
\name{TRR_simulated_data}
\alias{TRR_simulated_data}
\title{Simulated tensor response regression data for a single subject}
\usage{
TRR_simulated_data(
  subjects = 1,
  n.time = 200,
  margin_sizes = c(20, 20),
  CNR = 1,
  k = 0.3,
  num_active_regions = 1,
  obs.var = 1
)
}
\arguments{
\item{subjects}{The number of subjects in the data (a scalar)}

\item{n.time}{The length of the tensor time series (a scalar)}

\item{margin_sizes}{The sizes of the tensor coefficient (a vector)}

\item{CNR}{The contrast-to-noise ratio (a scalar), as defined in Welvaert, M., & Rosseel, Y. (2013). On the definition of signal-to-noise ratio and contrast-to-noise ratio for fMRI data. PloS one, 8(11), e77089.}

\item{k}{The AR(1) autoregression coefficient (a scalar)}

\item{num_active_regions}{The number of nonzero regions in the tensor coefficient}

\item{obs.var}{The observation variance (a scalar)}
}
\value{
A list with objects \code{Y} (the tensor-valued response), \code{x} (the matrix-valued covariate), \code{true_betas} (the array of true tensor coefficient values), and \code{true_k} the true value for the autocorrelation coefficient.
}
\description{
Simulated tensor response regression data for a single subject
}
\examples{
input <- TRR_simulated_data()
}
