% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/400_BTRR_single_subject.R
\name{BTRR_single_subject}
\alias{BTRR_single_subject}
\title{Single-subject Bayesian Tensor Response Regression}
\usage{
BTRR_single_subject(
  input,
  n.iter = 100,
  n.burn = 0,
  ranks = 1,
  hyperparameters = NULL,
  save_after = NULL,
  save_llik = TRUE,
  save_dir = "."
)
}
\arguments{
\item{input}{an object of class \code{TRR_data} or a list with elements \code{Y} (an array with dimensions \eqn{p_1\times \cdots \times p_D \times T \times 1}) and \code{x} (a \eqn{T \times 1} matrix).}

\item{n.iter}{(a scalar) the number of posterior samples desired}

\item{n.burn}{(a scalar) the number of posterior samples to discard as a burn-in}

\item{ranks}{(a positive integer) the number of ranks in the PARAFAC/CP tensor decomposition}

\item{hyperparameters}{a list with the (scalar) elements \code{a.tau}, \code{b.tau}, \code{a.lambda}, \code{b.lambda}, \code{a.epsilon}, and \code{b.epsilon} defining the values of the hyperparameters within the model. If \code{NULL}, then default values will be used.}

\item{save_after}{(an integer) An .rds file will be saved every \code{save_after} MCMC iterations with all of the results to that point (helpful for getting intermediate results in less time)}

\item{save_llik}{(a logical) Should the log-likelihood be calculated and saved at each iteration? Doing so comes at a cost, but can be used for model diagnostics. Defaults to \code{TRUE}.}

\item{save_dir}{(a character) A path to a directory in which the temporary results will be saved. Defaults to the current working directory.}
}
\value{
A list with the posterior samples
}
\description{
Performs the MCMC to draw from the posterior distribution for the model published by Guhaniyogi and Spencer (2020).
}
\examples{
# Not run
# input <- TRR_simulated_data()
# results <- BTRR_single_subject(input)
}
