% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/900_BTRR_sampling_fns.R
\name{DIC}
\alias{DIC}
\title{Deviance Information Criterion}
\usage{
DIC(log_likelihood, burn_in = 0)
}
\arguments{
\item{log_likelihood}{A vector of the log-likelihood values}

\item{burn_in}{A scalar indicating how many of the values should be removed from the beginning of the \code{log_likelihood} vector}
}
\value{
The DIC approximation value (a scalar)
}
\description{
Calculate the Deviance Information Criterion (DIC) through an approximation
  based solely on the values of the log-likelihood at each iteration of the
  Markov Chain Monte Carlo results.
}
