% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/404_FTR_CP.R
\name{FTR_CP}
\alias{FTR_CP}
\title{Frequentist Tensor Regression with the CP decomposition}
\usage{
FTR_CP(y, X, eta, rank = 1, epsilon = 1e-08)
}
\arguments{
\item{y}{a vector of response values}

\item{X}{an array of covariate values}

\item{eta}{a matrix of nuisance covariates}

\item{rank}{The rank of the CP decomposition to be used}

\item{epsilon}{a value for the stopping rule of the algorithm. Specifically,
this is the upper bound for the differences in the log-likelihood between
two iterations of the algorithm.}
}
\value{
A list with elements \code{gam} (vector coefficient result),
  \code{betas} (tensor decomposition components), \code{B} (the tensor ,
  coefficient), and \code{total_time} (time spent to complete the analysis).
}
\description{
Frequentist Tensor Regression with the CP decomposition
}
\examples{
\dontrun{
input <- TR_simulated_data()
results <- FTR_CP(input$y, input$X, input$eta)
}
}
