% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/906_ftr_functions.R
\name{ftr_log_likelihood}
\alias{ftr_log_likelihood}
\title{Obtain the log-likelihood for the tensor regression linear model}
\usage{
ftr_log_likelihood(y, X, eta, B, gam)
}
\arguments{
\item{y}{vector}

\item{X}{tensor covariate}

\item{eta}{design matrix}

\item{B}{tensor coefficient}

\item{gam}{vector coefficients}
}
\value{
scalar
}
\description{
Obtain the log-likelihood for the tensor regression linear model
}
\keyword{internal}
