% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/405_FTRTucker.R
\name{FTRTucker}
\alias{FTRTucker}
\title{Frequentist tensor regression with the Tucker decomposition}
\usage{
FTRTucker(
  input,
  ranks = NULL,
  epsilon = 1e-04,
  betas_LASSO = FALSE,
  G_LASSO = TRUE,
  step_limit = 1000
)
}
\arguments{
\item{input}{An object of class \code{TR_data} that contains (at least) the
elements \code{y} (a vector of response values) and \code{X} (an array of
covariate values). Optionally, \code{eta} (a matrix of nuisance covariates)
can also be included. Other list elements will be ignored.}

\item{ranks}{The ranks to be used with the Tucker decomposition. This should
be a vector with the same length as the tensor covariate for each subject.}

\item{epsilon}{a value for the stopping rule of the algorithm. Specifically,
this is the upper bound for the differences in the log-likelihood between
two iterations of the algorithm.}

\item{betas_LASSO}{(logical) Should the LASSO be applied to the betas in the
Tucker tensor decomposition? Defaults to \code{FALSE}.}

\item{G_LASSO}{(logical) Should the LASSO be applied to the core tensor in the
Tucker tensor decomposition? Defaults to \code{TRUE}.}

\item{step_limit}{The maximum number of steps that can be taken before
deciding that the algorithm did not converge}
}
\value{
A list with elements \code{gam} (vector coefficient result),
  \code{betas} (tensor decomposition components), \code{G} (the core tensor in the tensor decomposition),
  \code{B} (the tensor coefficient), \code{llik} (the value of the
  log-likelihood) and \code{total_time} (time spent to complete the analysis).
}
\description{
Frequentist tensor regression with the Tucker decomposition
}
\examples{
\dontrun{
input <- TR_simulated_data()
results <- FTRTucker(input)
}
}
