% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/901_BTRR_GGM_sampling_fns.R
\name{stick_break_log_posterior}
\alias{stick_break_log_posterior}
\title{Get the log posterior under the stick-breaking prior for BTRR_GGM}
\usage{
stick_break_log_posterior(Xi_g, current_rank, BtWB_g, p_g, tau_g, alpha_g)
}
\arguments{
\item{Xi_g}{region sticks}

\item{current_rank}{rank through the cycle}

\item{tau_g}{region tau}

\item{alpha_g}{region alpha}

\item{betas_g}{region beta}

\item{omega_g}{region omega}
}
\value{
The value for the log of the posterior density under the stick
breaking prior
}
\description{
Get the log posterior under the stick-breaking prior for BTRR_GGM
}
\keyword{internal}
